Move "sse" to "lib".
