+++ pylib-v1/ailets/models/gpt4o/response_to_messages.py
@@ -18,7 +18,7 @@ from ailets.cons.util import (
     iter_streams_objects,
     write_all,
 )
-from ailets.models.gpt4o.sse import SseHandler, is_sse_object
+from ailets.models.gpt4o.lib.sse import SseHandler, is_sse_object
 from ailets.models.gpt4o.lib.typing import (
     Gpt4oMessage,
     is_gpt4o_image,
